fix(logic): unblock #2847 quality lint (work-target constant + file size) (Refs #2847)
Use kWorkTargetExplore instead of a raw "explore" literal in the OW
feedstock prospect-localization test (repo.work_target_constants), and
extract the new ownsProspectedOldWorldMineralFeedstockTile /
hasIdleExplorerUnit predicates into
full_ai_civilian_work_ow_feedstock_localization.dart so
full_ai_civilian_work_selection.dart returns under the 1000 non-comment
line limit (repo.dart_file_non_comment_line_size). Predicates remain
exported via ai_api.dart; behavior unchanged.
