+++ packages/colonizethis_logic/lib/ai_api.dart
@@ -5,15 +5,15 @@
 library;
 
 export 'src/ai/ai_control.dart' show isAiControlled;
+export 'src/ai/full_ai_civilian_work_ow_feedstock_localization.dart'
+    show hasIdleExplorerUnit, ownsProspectedOldWorldMineralFeedstockTile;
 export 'src/ai/full_ai_civilian_work_selection.dart'
     show
         FullAiCivilianWorkIdle,
         FullAiCivilianWorkSelectionResult,
         feedstockBootstrapBuildImprovementCastIronWaived,
         feedstockExtractionResourceIdsForPlayer,
-        hasIdleExplorerUnit,
         kRegimentBuildInputFeedstockExtractionScoreBoost,
-        ownsProspectedOldWorldMineralFeedstockTile,
         regimentBuildInputFeedstockExtractionResourceIds,
         regimentBuildInputFeedstockImprovementInputCost,
         selectFullAiCivilianWorkOrders,
